fix: boxing of int32 values and enhance tests
- Updated BoxI32 function to correctly box int32 values by converting to uint32 before boxing.
- Enhanced TestBoxI32 to verify the boxed value's kind and ensure proper unboxing.
- Added IsNumeric method to Kind type to check if a kind is numeric (i32, i64, f32, f64).

+++ types/boxed.go
@@ -31,7 +31,7 @@ func IsBoxable(v int64) bool {
 }
 
 func BoxI32(v int32) Boxed {
-	return Box(uint64(v), KindI32)
+	return Box(uint64(uint32(v)), KindI32)
 }
 
 func BoxI64(v int64) Boxed {
